+++ import-service/utils.ts
@@ -1,0 +1,1 @@
+export const S3_REGION = 'eu-west-1';
