--- product-service/utils.ts
@@ -1,7 +1,0 @@
-import { PGTransactionEnum } from './types';
-
-export const PGTransaction: PGTransactionEnum = {
-  begin: 'BEGIN',
-  commit: 'COMMIT',
-  rollback: 'ROLLBACK',
-};
